% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-helpers.R
\name{cat_info_sep}
\alias{cat_info_sep}
\title{Print an informational message with a preceding blank line}
\usage{
cat_info_sep(msg)
}
\arguments{
\item{msg}{The message to print.}
}
\value{
The value returned by \code{\link[cli:cli_abort]{cli::cli_inform()}}
}
\description{
Print an informational message with a preceding blank line
}
\examples{
cat_info_sep("Doing something")

}
