% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-helpers.R
\name{cat_rule_sep}
\alias{cat_rule_sep}
\title{Print an informational 'separator' rule with a preceding blank line}
\usage{
cat_rule_sep(title = NULL, width = NULL)
}
\arguments{
\item{title}{The title to print for the rule, or NULL for none. Default: NULL}

\item{width}{The rule width, or NULL for half the console width. Default:
NULL}
}
\value{
The value returned by \code{\link[cli:cat_line]{cli::cat_rule()}}
}
\description{
Print an informational 'separator' rule with a preceding blank line
}
\examples{
cat_rule_sep()
cat_rule_sep("An informational message")
cat_rule_sep("An informational message", width = cli::console_width())

}
