% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4a_CollateResults_July_2021.R
\name{collate_results_jul_2021}
\alias{collate_results_jul_2021}
\title{Collate results for July 2021}
\usage{
collate_results_jul_2021(
  config = NULL,
  config_overrides = NULL,
  msoa_lim = NULL
)
}
\arguments{
\item{config}{A list containing parameters to control the run.
TODO: Describe the config options and their defaults}

\item{config_overrides}{A list containing settings with which to override the
default config values if required.}

\item{msoa_lim}{Number of MSOAs to process. Default: Number of unique MSOAs}
}
\value{
NULL (invisibly).
}
\description{
Collate results for July 2021
}
\examples{
\dontrun{
collate_results_jul_2021()
}

}
