% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimex-package.R
\docType{package}
\name{dimex-package}
\alias{dimex}
\alias{dimex-package}
\title{dimex: Data Integration Model for personal EXposures (DIMEX)}
\description{
Models personal exposure to pollution. Package derived from the SPFFinalReport repository (created for the case study and report for the final SPF reporting period).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://UoMResearchIT.github.io/SPFFinalReport}
  \item \url{https://github.com/UoMResearchIT/SPFFinalReport}
}

}
\author{
\strong{Maintainer}: Toni Price \email{maintainer-placeholder@example.com}

Authors:
\itemize{
  \item Matthew Thomas
}

}
\keyword{internal}
