% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_Activities_2021.R
\name{drop_locn}
\alias{drop_locn}
\title{Drop the name of the 'location' column from a vector of names}
\usage{
drop_locn(str_nms, nm_to_drop = NULL)
}
\arguments{
\item{str_nms}{A character vector of names.}

\item{nm_to_drop}{A character string: the name to drop from \code{str_nms}.}
}
\value{
The supplied character vector excluding \code{nm_to_drop}.
}
\description{
Drop the name of the 'location' column from a vector of names
}
\examples{
drop_locn(c("age", "type", "activity", "location"))
# => c("age", "type", "activity")

drop_locn(c("age", "type", "activity", "locn"), nm_to_drop = "locn")
# => c("age", "type", "activity")

}
