% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-sys-env-vars.R
\name{ensure_sys_env_vars}
\alias{ensure_sys_env_vars}
\title{Ensure the presence of required system environment variables}
\usage{
ensure_sys_env_vars(varnames = NULL)
}
\arguments{
\item{varnames}{A character vector of variable names to check for in the
user's environment.}
}
\value{
NULL (invisibly)
}
\description{
Check for presence of required system environment variables and abort if they
are not found.
}
\examples{
\dontrun{
ensure_sys_env_vars()
}

}
