% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-validation.R
\name{ensure_valid_env}
\alias{ensure_valid_env}
\title{Check the validity of a specified 'environment' string}
\usage{
ensure_valid_env(env, null_ok = NULL)
}
\arguments{
\item{env}{A character string specifying the environment for which to get the
path. This should be one of 'main', 'ref' or 'test'. See section \emph{Details}
for more information.}

\item{null_ok}{Whether it is valid for \code{env} to be NULL.}
}
\value{
NULL (invisibly)
}
\description{
Check the validity of a specified 'environment' string
}
\details{
Valid environments for data are: \if{html}{\out{<br>}}
\itemize{
\item main: Data produced by running the package code
\item ref:  Reference data (historical data which is used for regression
testing)
\item test: Test data
}
}
\examples{
ensure_valid_env("ref")
\dontrun{
ensure_valid_env("nonexistent") # => Causes an error
}

}
