% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-validation.R
\name{ensure_valid_phase}
\alias{ensure_valid_phase}
\title{Check the validity of a specified 'phase' string}
\usage{
ensure_valid_phase(phase, null_ok = NULL)
}
\arguments{
\item{phase}{A character string specifying the phase. Valid options are:
raw, processed and output}

\item{null_ok}{Whether it is valid for \code{phase} to be NULL.}
}
\value{
NULL (invisibly)
}
\description{
Check the validity of a specified 'phase' string
}
\examples{
ensure_valid_phase("output")
\dontrun{
ensure_valid_phase("nonexistent") # => Causes an error
}

}
