% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helpers.R
\name{env_should_run_hl_data_test}
\alias{env_should_run_hl_data_test}
\title{Check whether to run a particular high-level data test}
\usage{
env_should_run_hl_data_test(env_var_id = NULL)
}
\arguments{
\item{env_var_id}{The system environment variable that specified whether or
not to run the corresponding test.}
}
\value{
A boolean: TRUE if either the specified test is configured to run or
all high-level data tests are configured to run, otherwise FALSE.
}
\description{
Check whether to run a particular high-level data test
}
\examples{
\dontrun{
env_should_run_hl_data_test()
env_should_run_hl_data_test("DIMEX_RUN_HLDT_1D")
}

}
