% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{escape_txt}
\alias{escape_txt}
\title{Escape special characters in the given string}
\usage{
escape_txt(txt)
}
\arguments{
\item{txt}{The string which should be escaped so that it can function as a
regular expression.}
}
\value{
The given string with regex metacharacters escaped.
}
\description{
Given a character string, ensure this would function as a regular expression
by escaping regex metacharacters which appear in the string.
}
\examples{
escape_txt(".") # => "\\."
escape_txt("ab[c]") # => "ab\\[c\\]"
escape_txt("$30") # => "\\$30"

}
