% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-template.R
\name{generate_cfg_template}
\alias{generate_cfg_template}
\title{Create a config template containing default values}
\usage{
generate_cfg_template()
}
\value{
A list of lists containing the required config keys and values.
}
\description{
Create a config template containing default values
}
\details{
TODO: Describe config format and values
}
\examples{
\dontrun{
generate_cfg_template()
}

}
