% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-msoa.R
\name{get_area_id_list}
\alias{get_area_id_list}
\title{Get a list of unique MSOA area id's, possibly subsetted}
\usage{
get_area_id_list(area_ids, msoa_lim = NULL)
}
\arguments{
\item{area_ids}{The area id's as a character vector.}

\item{msoa_lim}{The maximum number of area id's to return, or NA for all of
them. Default: NA}
}
\value{
The sorted unique area id's, limited to \code{msoa_lim} if it is supplied.
}
\description{
Retrieve a sorted list of unique Middle layer Super Output Area id's, limited
to a maximum of \code{msoa_lim} if it is provided.
}
\examples{
\dontrun{
pop_dat <- get_pop_dat()
# Get the first 5 unique area id's
get_area_id_list(pop_dat$area_id, 5)
}
}
