% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_cfg}
\alias{get_cfg}
\title{Get a configuration to use for a run.}
\usage{
get_cfg(cfg = NULL, overrides = NULL, cfg_dir = NULL, cfg_name = NULL)
}
\arguments{
\item{cfg}{A list of nested lists, with 'leaf' nodes specifying configuration
values. Default: The template config returned by \code{\link[=generate_cfg_template]{generate_cfg_template()}}}

\item{overrides}{A list containing settings with which to override the
defaults. See section \emph{Details} for more information.}

\item{cfg_dir}{A character string specifying the config directory, or NULL
for the default. Default: An OS-specific path for user configuration files
in the user's home directory.}

\item{cfg_name}{A character string specifying the config file name, or NULL
for the default. Default: 'config.yml'}
}
\value{
A list containing settings for a run, constructed from defaults which
would be overridden if \code{config} and/or \code{overrides} are supplied. See
section \emph{Details} for more information.
}
\description{
Get a configuration to use for a run.
}
\details{
If \code{cfg} is supplied, this will be the returned config unless
\code{overrides} is also supplied, in which case the values in \code{overrides} will
be substituted within \code{cfg}. If both \code{cfg} and \code{overrides} are NULL then
the returned config will be the saved user config. If \code{cfg} is NULL and
\code{overrides} is supplied, the returned config will be the saved user config
with the values in \code{overrides} in overrides substituted.
}
\examples{
get_cfg()
get_cfg(overrides = list(run = list(seed_val = 200)))

}
