% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-template.R
\name{get_cfg_template_keys}
\alias{get_cfg_template_keys}
\title{Retrieve a character vector of expected config keys}
\usage{
get_cfg_template_keys()
}
\value{
A character vector of expected config keys.
}
\description{
Retrieve a character vector of expected config keys
}
\details{
The returned keys should be the same as those returned by
\code{get_cfg_keys(cfg = generate_cfg_template())}.
}
\examples{
get_cfg_template_keys()

}
