% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-template.R
\name{get_cfg_template_name}
\alias{get_cfg_template_name}
\title{Retrieve the config template file name}
\usage{
get_cfg_template_name(cfg_template_name = NULL)
}
\arguments{
\item{cfg_template_name}{A character string specifying the template config
file name, or NULL for the default. Default: 'config-template.yml'}
}
\value{
A character string: the supplied config template name or the default
config template name if none was provided.
}
\description{
Retrieve the config template file name
}
\examples{
get_cfg_template_name()
get_cfg_template_name("conftemplate.yml")

}
