% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-template.R
\name{get_cfg_template_path}
\alias{get_cfg_template_path}
\title{Retrieve the config template path}
\usage{
get_cfg_template_path(template_dir_name = NULL, template_name = NULL)
}
\arguments{
\item{template_dir_name}{A character string: the name of the config template
sub-directory or NULL for the default. Default: 'config'}

\item{template_name}{A character string: the name of the config template file
or NULL for the default. Default: The name returned by
\code{\link[=get_cfg_template_name]{get_cfg_template_name()}}}
}
\value{
A character string: the path to the config template file.
}
\description{
Retrieve the config template path
}
\examples{
get_cfg_template_path()
\dontrun{
get_cfg_template_path("cfg")
get_cfg_template_path(template_name = "cfg-file.yml")
get_cfg_template_path("cfg", "cfg-file.yml")
}

}
