% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_cfg_val}
\alias{get_cfg_val}
\title{Get a config value from the config}
\usage{
get_cfg_val(key, cfg_dir = NULL, cfg_name = NULL, cfg = NULL)
}
\arguments{
\item{key}{A single character string of hierarchical keys to look up with
each level of the hierarchy separated by a dot.}

\item{cfg_dir}{A character string specifying the config directory, or NULL
for the default. Default: An OS-specific path for user configuration files
in the user's home directory.}

\item{cfg_name}{A character string specifying the config file name, or NULL
for the default. Default: 'config.yml'}

\item{cfg}{A list of nested lists, with 'leaf' nodes specifying configuration
values. Default: The template config returned by \code{\link[=generate_cfg_template]{generate_cfg_template()}}}
}
\value{
A list of the requested value retrieved from the config. If the key
specification corresponds to a 'leaf' entry, this will be returned as a
list of length 1; if it corresponds to a list, this will be returned as a
list of length > 1.
}
\description{
Retrieve a config value from the user config or the supplied config.
}
\examples{
# This is not a 'leaf' entry so returns a list of length > 1
x <- get_cfg_val("store.dat.raw")

# This is a 'leaf' entry so returns a list of length 1 (character)
x <- get_cfg_val("store.dat.raw.base_dir", cfg = generate_cfg_template())

}
