% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{get_config}
\alias{get_config}
\title{Create a config object to specify parameters for a run}
\usage{
get_config(config = NULL, overrides = NULL)
}
\arguments{
\item{config}{A list containing parameters to control the run.
TODO: Describe the config options and their defaults}

\item{overrides}{A list containing settings with which to override the
defaults. If NULL then the returned list will contain all default values.}
}
\value{
A list containing settings for a run, constructed from defaults which
would be overridden if \code{config} and/or \code{overrides} are supplied.
}
\description{
Create a config object to specify parameters for a run
}
\examples{
get_config()
get_config(overrides = list(seed_val = 200))

}
