% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_Activities_2021.R
\name{get_grouping_vars}
\alias{get_grouping_vars}
\title{Get grouping variables for the time use survey data}
\usage{
get_grouping_vars(strata = NULL)
}
\arguments{
\item{strata}{An optional character vector containing names of stratification
variables. If supplied, these are added to the activity and location
variables to be used for grouping. Default: none.}
}
\value{
A character vector of names of grouping variables. These will be
'activity' and 'location' together with any stratification variables (if
supplied).
}
\description{
Get grouping variables for the time use survey data
}
\examples{
get_grouping_vars()
get_grouping_vars(c("sex", "agegr4", "nssec5", "daytype"))

}
