% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-validation.R
\name{get_invalid_cfg_keys}
\alias{get_invalid_cfg_keys}
\title{Retrieve any invalid keys from a config}
\usage{
get_invalid_cfg_keys(
  exp_keys = NULL,
  cfg_dir = NULL,
  cfg_name = NULL,
  cfg = NULL
)
}
\arguments{
\item{exp_keys}{A character vector of expected config keys or NULL for the
default. Default: Keys as returned by \code{\link[=get_cfg_template_keys]{get_cfg_template_keys()}}}

\item{cfg_dir}{A character string specifying the config directory, or NULL
for the default. Default: An OS-specific path for user configuration files
in the user's home directory.}

\item{cfg_name}{A character string specifying the config file name, or NULL
for the default. Default: 'config.yml'}

\item{cfg}{A list of nested lists, with 'leaf' nodes specifying configuration
values. Default: The template config returned by \code{\link[=generate_cfg_template]{generate_cfg_template()}}}
}
\value{
A character vector containing any invalid keys found in the config.
}
\description{
Identifies keys in the config which do not exist in \code{exp_keys}.
}
\examples{
get_invalid_cfg_keys()

}
