% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_Activities_2021.R
\name{get_most_popular_locn}
\alias{get_most_popular_locn}
\title{Get most popular location for each activity}
\usage{
get_most_popular_locn(tus_dat, col_nm, missing_ids, strata = NULL)
}
\arguments{
\item{tus_dat}{A data frame of the time use survey data. This is expected to
be in the format saved by \code{\link[=run_data_prep_tus]{run_data_prep_tus()}}.}

\item{col_nm}{A character string: name of new column to create which will
hold the most popular location for each activity (possibly by strata).}

\item{missing_ids}{An integer vector of the numeric id's which indicate
missing locations.}

\item{strata}{An optional character vector containing names of stratification
variables. If supplied, these are added to the activity and location
variables to be used for grouping. Default: none.}
}
\value{
The summarised time use survey data with 'most popular' location
added and rows then filtered to include one line for each combination of
strata and activity. If there are multiple locations matching the maximum
proportion of time spent in each location by stratum and activity then the
first location id will be chosen.
}
\description{
Get most popular location for each activity, possibly with stratification.
}
\examples{
\dontrun{
tus_dat <- get_tus_dat()
missing_ids <- c(-9, 0, 10, 90)
strata <- c("sex", "agegr4", "nssec5", "daytype")
get_most_popular_locn(tus_dat, "most_popular_locn", missing_ids, strata)
}
}
