% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-path.R
\name{get_path_components}
\alias{get_path_components}
\title{Get path components from a path specification}
\usage{
get_path_components(path)
}
\arguments{
\item{path}{The path from which to separate out components. This can be with
unix-style or windows-style separators.}
}
\value{
A character vector of the path components from the given path.
}
\description{
Split a path into its components in a platform-independent way. These
components can then be used for reconstructing the path on a different
platform.
}
\examples{
linux_path_1 <- "/path/to-somewhere/else"
linux_path_2 <- "path/to somewhere/else"
linux_path_3 <- "/path/to somewhere/else"
windows_path_1 <- "C:\\\\Path\\\\to somewhere\\\\else"
windows_path_2 <- "Path\\\\to somewhere\\\\else"

}
