% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_pkg_cfg_dir}
\alias{get_pkg_cfg_dir}
\title{Get the package config directory}
\usage{
get_pkg_cfg_dir(template_dir_name = NULL)
}
\arguments{
\item{template_dir_name}{A character string: the name of the config template
sub-directory or NULL for the default. Default: 'config'}
}
\value{
A character string: the path to the package config template
directory.
}
\description{
Retrieve the package config directory which is the directory for storing a
config template. If there is no pre-existing user config, this template gets
copied to the user config by \code{\link[=write_user_cfg]{write_user_cfg()}}.
}
\examples{
get_pkg_cfg_dir()
\dontrun{
get_pkg_cfg_dir("cfg")
}

}
\seealso{
\code{\link[=write_user_cfg]{write_user_cfg()}}
}
