% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_pop_dat}
\alias{get_pop_dat}
\title{Get the wrangled population data}
\usage{
get_pop_dat(env = NULL, cfg_dir = NULL, cfg_name = NULL, cfg = NULL)
}
\arguments{
\item{env}{A character string specifying the environment for which to get the
path. This should be one of 'main', 'ref' or 'test'.}

\item{cfg_dir}{A character string specifying the config directory, or NULL
for the default. Default: An OS-specific path for user configuration files
in the user's home directory.}

\item{cfg_name}{A character string specifying the config file name, or NULL
for the default. Default: 'config.yml'}

\item{cfg}{A list of nested lists, with 'leaf' nodes specifying configuration
values. Default: The template config returned by \code{\link[=generate_cfg_template]{generate_cfg_template()}}}
}
\value{
A \code{data.frame} containing the population data. See \code{\link[=data.frame]{data.frame()}}.
}
\description{
Get the wrangled population data
}
\examples{
\dontrun{
get_pop_dat()
}

}
