% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_root}
\alias{get_root}
\title{Get the root path for an environment}
\usage{
get_root(env = NULL)
}
\arguments{
\item{env}{A character string specifying the environment for which to get the
path. This should be one of 'main', 'ref' or 'test'. See section \emph{Details}
for more information.}
}
\value{
A character string: the root path for the given environment.
}
\description{
Retrieve the root (base) path for the given environment.
}
\details{
TODO: Add details for setting root paths according to env \if{html}{\out{<br>}}
\itemize{
\item DIMEX_STORE
\item DIMEX_STORE_REF
}
}
\examples{
get_root()
get_root("main")
get_root("ref")

}
