% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helpers.R
\name{get_stratified_samp}
\alias{get_stratified_samp}
\title{Create a small stratified sample from the given data}
\usage{
get_stratified_samp(
  .data,
  strata,
  n_per_stratum = NULL,
  replace = NULL,
  seed_val = NULL
)
}
\arguments{
\item{.data}{A data frame or data frame extension (e.g. a tibble).}

\item{strata}{A character vector of variables (column names) on which to
stratify.}

\item{n_per_stratum}{Sample size per stratum. Default: 3}

\item{replace}{A boolean: whether to sample with replacement or not. Default:
FALSE}

\item{seed_val}{An integer for setting the RNG seed. Default: 123}
}
\value{
The stratified sample created from the given data frame.
}
\description{
Create a small stratified sample from the given data
}
\examples{
df <- data.frame(
  group = rep(c("A", "B", "C"), each = 4),
  subgroup = rep(c("X", "Y"), times = 6),
  value = 1:24
)
df \%>\% get_stratified_samp(c("group", "subgroup"), n_per_stratum = 2)

\dontrun{
get_tus_dat() \%>\% get_stratified_samp()
}

}
