% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helpers.R
\name{get_test_cfg}
\alias{get_test_cfg}
\title{Test helper to create a test config}
\usage{
get_test_cfg()
}
\value{
A list which conforms to the package config structure
}
\description{
Test helper to create a test config
}
\examples{
\dontrun{
get_test_cfg()
}

}
