% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_user_cfg_dir}
\alias{get_user_cfg_dir}
\title{Retrieve the user-specific config dir}
\usage{
get_user_cfg_dir(cfg_dir = NULL)
}
\arguments{
\item{cfg_dir}{A character string specifying the config directory, or NULL
for the default. Default: An OS-specific path for user configuration files
in the user's home directory.}
}
\value{
A character string: the supplied config dir or the default config
dir if none was provided.
}
\description{
Retrieve the user-specific config dir
}
\examples{
get_user_cfg_dir()
get_user_cfg_dir(file.path("config"))

}
