% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_user_cfg_name}
\alias{get_user_cfg_name}
\title{Retrieve the user-specific config file name}
\usage{
get_user_cfg_name(cfg_name = NULL)
}
\arguments{
\item{cfg_name}{A character string specifying the config file name, or NULL
for the default. Default: 'config.yml'}
}
\value{
A character string: the supplied config name or the default config
name if none was provided.
}
\description{
Retrieve the user-specific config file name
}
\examples{
get_user_cfg_name()
get_user_cfg_name("conf.yml")

}
