% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_user_cfg_path}
\alias{get_user_cfg_path}
\title{Retrieve the config file path}
\usage{
get_user_cfg_path(cfg_dir = NULL, cfg_name = NULL)
}
\arguments{
\item{cfg_dir}{A character string specifying the config directory, or NULL
for the default. Default: An OS-specific path for user configuration files
in the user's home directory.}

\item{cfg_name}{A character string specifying the config file name, or NULL
for the default. Default: 'config.yml'}
}
\value{
A character string: the path to the config file, constructed from
\code{cfg_dir} and \code{cfg_name} or defaults if these are not suplied.
}
\description{
Retrieve the config file path
}
\examples{
get_user_cfg_path()
get_user_cfg_path(file.path("conf"), "configuration.yml")

}
