% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Establish whether particular values exist in a vector.}
\usage{
x \%notin\% y
}
\arguments{
\item{x}{A vector or NULL: the values to be matched.}

\item{y}{A vector or NULL: the values to be checked for no match.}
}
\value{
A logical vector, indicating if a match was \emph{not} located for each
element of x (i.e., the values are TRUE or FALSE).
}
\description{
This is the negation of the \code{\%in\%} operator.
}
\examples{
"a" \%notin\% c("b", "c")

}
\seealso{
\code{\link[=\%in\%]{\%in\%()}}
}
