% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_Activities_2021.R
\name{impute_missing}
\alias{impute_missing}
\title{Impute missing location data}
\usage{
impute_missing(.data, tus_dat, col_nm, missing_ids, strata = NULL)
}
\arguments{
\item{.data}{A data frame or data frame extension (e.g. a tibble).}

\item{tus_dat}{A data frame of the time use survey data. This is expected to
be in the format saved by \code{\link[=run_data_prep_tus]{run_data_prep_tus()}}.}

\item{col_nm}{A character string: name of new column to create which will
hold the most popular location for each activity (possibly by strata).}

\item{missing_ids}{An integer vector of the numeric id's which indicate
missing locations.}

\item{strata}{An optional character vector containing names of stratification
variables. If supplied, these are added to the activity and location
variables to be used for grouping. Default: none.}
}
\value{
The supplied \code{tus_dat} data frame with imputed data for missing
locations. Imputed values are computed as the most popular location for
activity by (optional) strata.
}
\description{
Impute missing location data
}
\examples{
\dontrun{
tus_dat <- get_tus_dat()
strata <- c("sex", "agegr4", "nssec5", "daytype")
impute_missing(tus_dat, tus_dat, "most_popular_locn", missing_ids, strata)
}

}
