% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_Activities_2021.R
\name{impute_tus_dat}
\alias{impute_tus_dat}
\title{Impute missing tus data}
\usage{
impute_tus_dat(env = NULL, cfg_dir = NULL, cfg_name = NULL, cfg = NULL)
}
\arguments{
\item{env}{A character string specifying the environment for which to get the
path. This should be one of 'main', 'ref' or 'test'.}

\item{cfg_dir}{A character string specifying the config directory, or NULL
for the default. Default: An OS-specific path for user configuration files
in the user's home directory.}

\item{cfg_name}{A character string specifying the config file name, or NULL
for the default. Default: 'config.yml'}

\item{cfg}{A list of nested lists, with 'leaf' nodes specifying configuration
values. Default: The template config returned by \code{\link[=generate_cfg_template]{generate_cfg_template()}}}
}
\value{
A data frame with the imputed tus data.
}
\description{
Impute missing tus data
}
\examples{
\dontrun{
# To override the config to save the imputed tus data:
overrides = list(store = list(save = list(imputed_tus_dat = TRUE)))
impute_tus_dat(cfg = get_cfg(overrides = overrides))
}
}
