% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{is_int_val}
\alias{is_int_val}
\title{Tests if a value is an integer.}
\usage{
is_int_val(x)
}
\arguments{
\item{x}{The value to test.}
}
\value{
TRUE if the value is a number with no decimal value, FALSE otherwise.
}
\description{
Tests if a value is an integer.
}
\examples{
is_int_val("5")     # => FALSE
is_int_val(10)      # => TRUE
is_int_val(16.5)    # => FALSE
is_int_val("apple") # => FALSE
is_int_val(-1)      # => TRUE

}
