% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helpers.R
\name{is_r_cmd_check}
\alias{is_r_cmd_check}
\title{Check if R CMD CHECK is currently running}
\usage{
is_r_cmd_check()
}
\value{
A boolean: TRUE if the system environment variable
'\emph{R_CHECK_PACKAGE_NAME}' is non-empty (which should indicate that
'R CMD CHECK' is currently running), otherwise FALSE.
}
\description{
Check if R CMD CHECK is currently running
}
\examples{
\dontrun{
is_r_cmd_check()
}

}
