% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{is_strict_vec}
\alias{is_strict_vec}
\title{Check if an object is a vector}
\usage{
is_strict_vec(x, null_ok = NULL)
}
\arguments{
\item{x}{The object to check.}

\item{null_ok}{A boolean indicating whether a NULL value for \code{x} is
acceptable. Default: FALSE}
}
\value{
If \code{null_ok} is TRUE, returns TRUE if \code{x} is NULL or \code{x} is a vector,
otherwise returns FALSE; if \code{null_ok} is FALSE, throws an error if \code{x} is
NULL and returns TRUE if \code{x} is a vector, otherwise returns FALSE.
}
\description{
Check if an object is a vector
}
\examples{
is_strict_vec(c(10, 20, 30)) # => TRUE
is_strict_vec(list(10, 20, 30)) # => FALSE

}
