% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helpers.R
\name{local_mk_dir}
\alias{local_mk_dir}
\title{Test helper to create a temporary directory}
\usage{
local_mk_dir(
  dir = fs::file_temp(pattern = "R-dimex-tst-dir_"),
  envir = parent.frame()
)
}
\arguments{
\item{dir}{A temporary directory as a character string or as would be
returned by \code{\link[fs:file_temp]{fs::file_temp()}}.}

\item{envir}{The environment to use for scoping which must default to
\code{\link[=parent.frame]{parent.frame()}} - do NOT changed this by supplying an alternative as this
is used by \code{\link[withr:defer]{withr::defer()}}.}
}
\value{
The name of the temporary directory as a character string (not as an
\code{fs_path} object).
}
\description{
Test helper to create a temporary directory
}
\examples{
\dontrun{
tmp_dir <- local_mk_dir()
}

}
\references{
\href{https://testthat.r-lib.org/articles/test-fixtures.html}{Test fixtures in testthat}
}
