% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helpers.R
\name{local_sys_env_vars}
\alias{local_sys_env_vars}
\title{Test helper to temporarily set sys env vars}
\usage{
local_sys_env_vars(vars, envir = parent.frame())
}
\arguments{
\item{vars}{A named character vector with names corresponding to vars that
should be temporarily changed with the values they should be changed to.}

\item{envir}{The environment to use for scoping which must default to
\code{\link[=parent.frame]{parent.frame()}} - do NOT changed this by supplying an alternative as this
is used by \code{\link[withr:defer]{withr::defer()}}.}
}
\value{
NULL (invisibly)
}
\description{
Test helper to temporarily set sys env vars
}
\details{
A call to \code{local_sys_env_vars()} will alter the value of the specified system
environment variables and set up a deferred event to set them back to their
original values when the calling function exits (i.e., within the
appropriate scope as controlled by the \code{envir} argument).
}
\examples{
\dontrun{
# For interactive use:
local_sys_env_vars(c(DIMEX_STORE = "xxx", DIMEX_STORE_REF = "yyy"))
# Run the deferred events to set the sys env vars back to their original
# values:
withr::deferred_run()
# Note: Use `withr::deferred_clear()` to clear the events without running
}

}
\references{
\href{https://testthat.r-lib.org/articles/test-fixtures.html}{Test fixtures in testthat}
}
