% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helpers.R
\name{local_write_cfg}
\alias{local_write_cfg}
\title{Test helper to write a test config to file}
\usage{
local_write_cfg(
  cfg_name = "tmp_cfg.yml",
  dir = fs::file_temp(pattern = "R-dimex-tst-cfg_"),
  envir = parent.frame()
)
}
\arguments{
\item{cfg_name}{A character string: name of the config file to write.}

\item{dir}{A temporary directory as a character string or as would be
returned by \code{\link[fs:file_temp]{fs::file_temp()}}.}

\item{envir}{The environment to use for scoping which must default to
\code{\link[=parent.frame]{parent.frame()}} - do NOT changed this by supplying an alternative as this
is used by \code{\link[withr:defer]{withr::defer()}}.}
}
\value{
The name of the temporary directory that contains the saved config
file as a character string (not as an \code{fs_path} object).
}
\description{
Test helper to write a test config to file
}
\details{
A call to \code{local_write_cfg()} will write a config to file to a temporary
directory and set up a deferred event to delete the directory (and file)
when the calling function exits (i.e., within the appropriate scope as
controlled by the \code{envir} argument).
}
\examples{
\dontrun{
tmp_dir <- local_write_cfg()
}

}
\references{
\href{https://testthat.r-lib.org/articles/test-fixtures.html}{Test fixtures in testthat}
}
