% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{parse_steps_to_run}
\alias{parse_steps_to_run}
\title{Parse which steps to run from a string}
\usage{
parse_steps_to_run(step_ids)
}
\arguments{
\item{step_ids}{A character string specifying comma-separated identifiers for
steps in the workflow. These can be:
1a, 1b, 1c, 1d, 1e, 1f, 2, 3a, 3b, 4a, 4b, 5a, 5b}
}
\value{
A character vector of the step IDs.
}
\description{
Parse which steps to run from a string
}
\examples{
parse_steps_to_run("1a")
parse_steps_to_run("1a, 1b")
parse_steps_to_run("1a,1b,1c,")

}
