% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{parse_string}
\alias{parse_string}
\title{Parse a string into a number, NA or to remain as a string}
\usage{
parse_string(x)
}
\arguments{
\item{x}{The character string to parse}
}
\value{
One of: a number, NA or the original string
}
\description{
Parse a string into a number, NA or to remain as a string
}
\examples{
parse_string("NA")
parse_string("notanumber")
parse_string("100")
parse_string("0.45")

}
