% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{path_from_cfg}
\alias{path_from_cfg}
\title{Get a path value from the configuration}
\usage{
path_from_cfg(env, key, cfg_dir = NULL, cfg_name = NULL, cfg = NULL)
}
\arguments{
\item{env}{A character string specifying the environment for which to get the
path. This should be one of 'ref', 'act' or 'test'. See section \emph{Details}
for more information.}

\item{key}{A single character string of hierarchical keys to look up with
each level of the hierarchy separated by the value of \code{key_sep} (which is a
dot by default).}

\item{cfg_dir}{A character string specifying the config directory, or NULL
for the default. Default: An OS-specific path for user configuration files
in the user's home directory.}

\item{cfg_name}{A character string specifying the config file name, or NULL
for the default. Default: 'config.yml'}

\item{cfg}{A list of nested lists, with 'leaf' nodes specifying configuration
values. Default: The template config returned by \code{\link[=generate_cfg_template]{generate_cfg_template()}}}
}
\value{
A character string: the requested path, retrieved from the default
config or the given config, if provided.
}
\description{
Get a path value from the configuration
}
\examples{
path_from_cfg("ref", "store.dat.wrangled.shapefiles.names_ref.msoa_file",
              cfg = generate_cfg_template())

}
