% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-path.R
\name{path_from_components}
\alias{path_from_components}
\title{Construct a file system path from individual path components}
\usage{
path_from_components(components)
}
\arguments{
\item{components}{A character vector: the path components.}
}
\value{
The constructed path.
}
\description{
Construct a file system path from individual path components
}
\examples{
\dontrun{
# On *nix/macOS:
path_from_components(c("", "an", "absolute", "path"))
path_from_components(c("a", "relative", "path"))
# On Windows:
path_from_components(c("\\\\\\\\", "servername", "sharename"))
path_from_components(c("C:", "Data"))
}

}
