% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5a_PlotResults_July_2021.R
\name{plot_results_jul_2021}
\alias{plot_results_jul_2021}
\title{Plot results for July 2021}
\usage{
plot_results_jul_2021()
}
\value{
NULL (invisibly).
}
\description{
Plot results for July 2021
}
\examples{
\dontrun{
plot_results_jul_2021()
}

}
