% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_Activities_2021.R
\name{process_activities}
\alias{process_activities}
\title{Process Activities data}
\usage{
process_activities(msoa_lim = NULL)
}
\arguments{
\item{msoa_lim}{Number of MSOAs to process. Default: Number of unique MSOAs}
}
\value{
NULL (invisibly).
}
\description{
Process Activities data
}
\examples{
\dontrun{
process_activities()
}

}
