% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{recurse_nested}
\alias{recurse_nested}
\title{Recursively apply a function to each element of a nested list}
\usage{
recurse_nested(x, func)
}
\arguments{
\item{x}{The element on which to operate, which could be a list or some other
type of object.}

\item{func}{The function to apply to each element of the list.}
}
\value{
A list which is the same as the list provided but with the given
function applied to each element.
}
\description{
Recursively apply a function to each element of a nested list
}
\examples{
nested <- list(
  a = "1",
  b = "12.5",
  c = "NA",
  d = "alphabet",
  e = list(
    first = "NA",
    second = "notanumber",
    third = "100",
    fourth = "0.45",
    fifth = list(
      one = "teal",
      two = "NA",
      three = "999"
    )
  )
)
recurse_nested(nested, parse_string)

}
