% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1c_DataPrep_TUS.R
\name{run_data_prep_tus}
\alias{run_data_prep_tus}
\title{Process the Time Use Survey data}
\usage{
run_data_prep_tus(cfg = NULL)
}
\arguments{
\item{cfg}{A list of nested lists, with 'leaf' nodes specifying configuration
values. Default: The template config returned by \code{\link[=generate_cfg_template]{generate_cfg_template()}}}
}
\value{
NULL (invisibly).
}
\description{
This function implements a data preparation step to process the Time Use
Survey data. It is useful only for its side effects, i.e. for saving the
processed data.
}
\examples{
\dontrun{
run_data_prep_tus()
}

}
