% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{run_workflow}
\alias{run_workflow}
\title{Run the steps in the data prep, analysis and visualisation workflow pipeline}
\usage{
run_workflow(
  steps_to_run = NULL,
  cfg = NULL,
  cfg_overrides = NULL,
  cfg_dir = NULL,
  cfg_name = NULL,
  banner_only = NULL
)
}
\arguments{
\item{steps_to_run}{A character string specifying comma-separated identifiers
for steps to run in the workflow. \if{html}{\out{<br>}}
These can be: \if{html}{\out{<br>}}
1a, 1b, 1c, 1d, 1e, 1f, 2, 3a, 3b, 4a, 4b, 5a, 5b \if{html}{\out{<br>}}
Default: "1a, 1b, 1c, 1d, 1e, 1f, 2, 3a, 3b, 4a, 4b, 5a, 5b" (run all
steps)}

\item{cfg}{A list of nested lists, with 'leaf' nodes specifying configuration
values. Default: The template config returned by \code{\link[=generate_cfg_template]{generate_cfg_template()}}}

\item{cfg_overrides}{An optional list containing settings with which to
override the config values for this run.}

\item{cfg_dir}{A character string specifying the config directory, or NULL
for the default. Default: An OS-specific path for user configuration files
in the user's home directory.}

\item{cfg_name}{A character string specifying the config file name, or NULL
for the default. Default: 'config.yml'}

\item{banner_only}{A boolean indicating whether to only show banners for each
step which would be run, or to run the function as well (useful for
checking which steps would be run).}
}
\value{
NULL (invisibly).
}
\description{
This function runs each step in the data preparation, data analysis and
visualisation pipeline. Note that it will generally take an appreciable
amount of time to run.
}
\examples{
run_workflow("1a", banner_only = TRUE)
\dontrun{
run_workflow("1a")
run_workflow("1a,1b,1c,1d,1e,1f,2,3a,3b,4a,4b,5a,5b", banner_only = TRUE)
run_workflow(banner_only = TRUE)
run_workflow("4a, 4b")
run_workflow("4a, 4b", banner_only = TRUE)
run_workflow("1a, 1b, 1c, 1d, 1e, 1f, 2, 3a, 3b, 4a, 4b")
run_workflow()
}

}
