% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-helpers.R
\name{show_workflow_banner}
\alias{show_workflow_banner}
\title{Print a banner to the console ahead of running a step in the workflow}
\usage{
show_workflow_banner(id, title, pad_width = NULL)
}
\arguments{
\item{id}{A character string for the step id.}

\item{title}{A character string which is the title of the step.}

\item{pad_width}{The width to which to pad the title with blanks, unless it
is NA in which case no padding will be done. Default: NA}
}
\value{
NULL (invisibly)
}
\description{
Print a banner to the console ahead of running a step in the workflow
}
\examples{
show_workflow_banner("1", "Step 1")
show_workflow_banner("1", "Step 1", 15)

}
