% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helpers.R
\name{skip_if_r_cmd_check_or_not_configured}
\alias{skip_if_r_cmd_check_or_not_configured}
\title{Skip a test under certain conditions}
\usage{
skip_if_r_cmd_check_or_not_configured(env_var_id = NULL)
}
\arguments{
\item{env_var_id}{The system environment variable that specified whether or
not to run the corresponding test.}
}
\value{
NULL (invisibly)
}
\description{
Skip a test if not running within the context of 'R CMD CHECK'  or not
configured to run, as determined by the value of the supplied system
environment variable.
}
\examples{
\dontrun{
skip_if_r_cmd_check_or_not_configured()
Sys.setenv(DIMEX_RUN_HLDT_1D = "false")
skip_if_r_cmd_check_or_not_configured("DIMEX_RUN_HLDT_1D")
}

}
