% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{sweep_key}
\alias{sweep_key}
\title{Ensure a hierarchical key string has the correct separator}
\usage{
sweep_key(key, key_sep = NULL, old_key_sep = NULL)
}
\arguments{
\item{key}{A single character string of hierarchical keys to look up with
each level of the hierarchy separated by the value of \code{key_sep} (which is a
dot by default).}

\item{key_sep}{A character string specifying the separator between
hierarchical key elements. Default: '.'}

\item{old_key_sep}{A character string specifying the separator between
hierarchical key elements that are used for \code{key}. Default: '.'}
}
\value{
The key with hierarchical elements separated by \code{key_sep}.
}
\description{
Ensure a hierarchical key string has the correct separator
}
\examples{
sweep_key("one.two.three") # => one.two.three
sweep_key("one.two.three", "#") # => one#two#three

}
