% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-template.R
\name{write_cfg_template}
\alias{write_cfg_template}
\title{Write the configuration template to file}
\usage{
write_cfg_template(template_dir_name = NULL, template_name = NULL, cfg = NULL)
}
\arguments{
\item{template_dir_name}{A character string: the name of the config template
sub-directory or NULL for the default. Default: 'config'}

\item{template_name}{A character string: the name of the config template file
or NULL for the default. Default: The name returned by
\code{\link[=get_cfg_template_name]{get_cfg_template_name()}}}

\item{cfg}{A list of nested lists, with 'leaf' nodes specifying configuration
values. Default: The template config returned by \code{\link[=generate_cfg_template]{generate_cfg_template()}}}
}
\value{
NULL (invisibly)
}
\description{
Write the configuration template to file
}
\examples{
write_cfg_template()

}
