% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4b_CollateResults_Q1_2021.R
\name{collate_results_q1_2021}
\alias{collate_results_q1_2021}
\title{Collate results for Q1 2021}
\usage{
collate_results_q1_2021(cfg = NULL, msoa_lim = NULL)
}
\arguments{
\item{cfg}{A list of nested lists, with 'leaf' nodes specifying configuration
values. Default: The template config returned by \code{\link[=generate_cfg_template]{generate_cfg_template()}}}

\item{msoa_lim}{Number of MSOAs to process. Default: Number of unique MSOAs}
}
\value{
NULL (invisibly).
}
\description{
Collate results for Q1 2021
}
\examples{
\dontrun{
collate_results_q1_2021()
}

}
