% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_cfg_pkg_id}
\alias{get_cfg_pkg_id}
\title{Get the package identifier for the user config location}
\usage{
get_cfg_pkg_id()
}
\value{
A character string: The package identifier for configuration
purposes.
}
\description{
Retrieve an identifier to use as a OS-specific directory name when creating
the package config. This will be used for creating an OS-specific path for
storing user configuration files in a user's home directory.
}
\examples{
get_cfg_pkg_id

}
