% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-msoa.R
\name{get_cfg_val_msoa}
\alias{get_cfg_val_msoa}
\title{Get the MSOA value from the config}
\usage{
get_cfg_val_msoa(key = NULL, cfg_dir = NULL, cfg_name = NULL, cfg = NULL)
}
\arguments{
\item{key}{A single character string of hierarchical keys to look up which is
flattened by separating each level of the hierarchy with a dot.}

\item{cfg_dir}{A character string specifying the config directory, or NULL
for the default. Default: An OS-specific path for user configuration files
in the user's home directory.}

\item{cfg_name}{A character string specifying the config file name, or NULL
for the default. Default: 'config.yml'}

\item{cfg}{A list of nested lists, with 'leaf' nodes specifying configuration
values. Default: The template config returned by \code{\link[=generate_cfg_template]{generate_cfg_template()}}}
}
\value{
An integer: the configured MSOA value which is either the number
specified, or, if this has a value of \code{NA}, the total number of MSOAs in
the given population data file.
}
\description{
Get the MSOA value from the config
}
\examples{
get_cfg_val_msoa(cfg = generate_cfg_template())

}
