% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5a_PlotResults_July_2021.R
\name{plot_results_jul_2021}
\alias{plot_results_jul_2021}
\title{Plot results for July 2021}
\usage{
plot_results_jul_2021(cfg = NULL)
}
\arguments{
\item{cfg}{A list of nested lists, with 'leaf' nodes specifying configuration
values. Default: The template config returned by \code{\link[=generate_cfg_template]{generate_cfg_template()}}}
}
\value{
NULL (invisibly).
}
\description{
Plot results for July 2021
}
\examples{
\dontrun{
plot_results_jul_2021()
}

}
