% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5a_PlotResults_July_2021.R
\name{plot_results_jul_2021}
\alias{plot_results_jul_2021}
\title{Plot results for July 2021}
\usage{
plot_results_jul_2021(config = NULL, config_overrides = NULL)
}
\arguments{
\item{config}{A list containing parameters to control the run.
TODO: Describe the config options and their defaults}

\item{config_overrides}{A list containing settings with which to override the
default config values if required.}
}
\value{
NULL (invisibly).
}
\description{
Plot results for July 2021
}
\examples{
\dontrun{
plot_results_jul_2021()
}

}
