% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5b_PlotResults_Q1_2021.R
\name{plot_results_q1_2021}
\alias{plot_results_q1_2021}
\title{Plot results for Q1 2021}
\usage{
plot_results_q1_2021()
}
\value{
NULL (invisibly).
}
\description{
Plot results for Q1 2021
}
\examples{
\dontrun{
plot_results_q1_2021()
}

}
