% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_Activities_2021.R
\name{process_activities}
\alias{process_activities}
\title{Process Activities data}
\usage{
process_activities(cfg = NULL, msoa_lim = NULL)
}
\arguments{
\item{cfg}{A list of nested lists, with 'leaf' nodes specifying configuration
values. Default: The template config returned by \code{\link[=generate_cfg_template]{generate_cfg_template()}}}

\item{msoa_lim}{Number of MSOAs to process. Default: Number of unique MSOAs}
}
\value{
NULL (invisibly).
}
\description{
Process Activities data
}
\examples{
\dontrun{
process_activities()
}

}
