% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3b_Exposures_Q1_2021.R
\name{process_exposures_q1_2021}
\alias{process_exposures_q1_2021}
\title{Process Exposures data for Q1 2021}
\usage{
process_exposures_q1_2021(cfg = NULL, msoa_lim = NULL)
}
\arguments{
\item{cfg}{A list of nested lists, with 'leaf' nodes specifying configuration
values. Default: The template config returned by \code{\link[=generate_cfg_template]{generate_cfg_template()}}}

\item{msoa_lim}{Number of MSOAs to process. Default: Number of unique MSOAs}
}
\value{
NULL (invisibly).
}
\description{
Process Exposures data for Q1 2021
}
\examples{
\dontrun{
process_exposures_q1_2021()
}

}
