% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{read_user_cfg}
\alias{read_user_cfg}
\title{Read the YAML config file into a named list}
\usage{
read_user_cfg(cfg_dir = NULL, cfg_name = NULL)
}
\arguments{
\item{cfg_dir}{A character string specifying the config directory, or NULL
for the default. Default: An OS-specific path for user configuration files
in the user's home directory.}

\item{cfg_name}{A character string specifying the config file name, or NULL
for the default. Default: 'config.yml'}
}
\value{
A named list of lists which corresponds to the loaded YAML config.
}
\description{
Read the YAML config file into a named list
}
\examples{
read_user_cfg()
\dontrun{
read_user_cfg(file.path("path", "to", "config"), "conf.yml")
}

}
