% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1d_DataPrep_PM25_CAMS.R
\name{run_data_prep_pm25_cams}
\alias{run_data_prep_pm25_cams}
\title{Process the PM25 CAMS data}
\usage{
run_data_prep_pm25_cams()
}
\value{
NULL (invisibly).
}
\description{
This function implements a data preparation step to process the PM25 CAMS
data. It is useful only for its side effects, i.e. for saving the processed
data.
}
\examples{
\dontrun{
run_data_prep_pm25_cams()
}

}
