% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1d_DataPrep_PM25_CAMS.R
\name{run_data_prep_pm25_cams}
\alias{run_data_prep_pm25_cams}
\title{Process the pm25 cams data}
\usage{
run_data_prep_pm25_cams(config = NULL, config_overrides = NULL)
}
\arguments{
\item{config}{A list containing parameters to control the run.
TODO: Describe the config options and their defaults}

\item{config_overrides}{A list containing settings with which to override the
default config values if required.}
}
\value{
NULL (invisibly).
}
\description{
This function implements a data preparation step to process the pm25 cams
data. It is useful only for its side effects, i.e. for saving the processed
data.
}
\examples{
\dontrun{
run_data_prep_pm25_cams()
}

}
