% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1e_DataPrep_PM25_EMEP.R
\name{run_data_prep_pm25_emep}
\alias{run_data_prep_pm25_emep}
\title{Process the PM25 EMEP data}
\usage{
run_data_prep_pm25_emep(cfg = NULL)
}
\arguments{
\item{cfg}{A list of nested lists, with 'leaf' nodes specifying configuration
values. Default: The template config returned by \code{\link[=generate_cfg_template]{generate_cfg_template()}}}
}
\value{
NULL (invisibly).
}
\description{
This function implements a data preparation step to process the PM25 EMEP
data. It is useful only for its side effects, i.e. for saving the processed
data.
}
\examples{
\dontrun{
run_data_prep_pm25_emep()
}

}
