% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1e_DataPrep_PM25_EMEP.R
\name{run_data_prep_pm25_emep}
\alias{run_data_prep_pm25_emep}
\title{Process the PM25 EMEP data}
\usage{
run_data_prep_pm25_emep(config = NULL, config_overrides = NULL)
}
\arguments{
\item{config}{A list containing parameters to control the run.
TODO: Describe the config options and their defaults}

\item{config_overrides}{A list containing settings with which to override the
default config values if required.}
}
\value{
NULL (invisibly).
}
\description{
This function implements a data preparation step to process the PM25 EMEP
data. It is useful only for its side effects, i.e. for saving the processed
data.
}
\examples{
\dontrun{
run_data_prep_pm25_emep()
}

}
