% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1f_DataPrep_PM25_GM.R
\name{run_data_prep_pm25_gm}
\alias{run_data_prep_pm25_gm}
\title{Process the PM25 GM data}
\usage{
run_data_prep_pm25_gm(
  config = NULL,
  config_overrides = NULL,
  use_cached_openair_data = NULL
)
}
\arguments{
\item{config}{A list containing parameters to control the run.
TODO: Describe the config options and their defaults}

\item{config_overrides}{A list containing settings with which to override the
default config values if required.}

\item{use_cached_openair_data}{TRUE if cached versions of the openair data
should be used, FALSE otherwise. Default: TRUE}
}
\value{
NULL (invisibly).
}
\description{
This function implements a data preparation step to process the PM25 GM data.
It is useful only for its side effects, i.e. for saving the processed data.
}
\examples{
\dontrun{
run_data_prep_pm25_gm()
}

}
