% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b_DataPrep_Population.R
\name{run_data_prep_population}
\alias{run_data_prep_population}
\title{Process the Population data}
\usage{
run_data_prep_population(config = NULL, config_overrides = NULL)
}
\arguments{
\item{config}{A list containing parameters to control the run.
TODO: Describe the config options and their defaults}

\item{config_overrides}{A list containing settings with which to override the
default config values if required.}
}
\value{
NULL (invisibly).
}
\description{
This function implements a data preparation step to process the Population
data. It is useful only for its side effects, i.e. for saving the processed
data.
}
\examples{
\dontrun{
run_data_prep_population()
}

}
