% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1a_DataPrep_StudyRegion.R
\name{run_data_prep_study_region}
\alias{run_data_prep_study_region}
\title{Process the Study Region data}
\usage{
run_data_prep_study_region(config = NULL, config_overrides = NULL)
}
\arguments{
\item{config}{A list containing parameters to control the run.
TODO: Describe the config options and their defaults}

\item{config_overrides}{A list containing settings with which to override the
default config values if required.}
}
\value{
NULL (invisibly).
}
\description{
This function implements a data preparation step to process the Study Region
data. It is useful only for its side effects, i.e. for saving the processed
data.
}
\examples{
\dontrun{
run_data_prep_study_region()
run_data_prep_study_region(config_overrides = list(seed_val = 10))
}

}
