% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1a_DataPrep_StudyRegion.R
\name{run_data_prep_study_region}
\alias{run_data_prep_study_region}
\title{Process the Study Region data}
\usage{
run_data_prep_study_region()
}
\value{
NULL (invisibly).
}
\description{
This function implements a data preparation step to process the Study Region
data. It is useful only for its side effects, i.e. for saving the processed
data.
}
\examples{
\dontrun{
run_data_prep_study_region()
}

}
